+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 - (**SystemTray**) Disable DorkBox update requests
+- (**Downloader**) Skip LocalSource downloading
 
 ### Fixed
 - (**Tracker**) Fix Shikimori